Move views tests to views sub-directory.
