+++ tests/views/adminstrate.py
